refactor: extract STA artifact publication into sta_artifacts module
Address review findings on the SDF publishing change:

- Move timing report/SDF/structured validation and copying out of the
  oversized module.py into a focused sta_artifacts module, leaving
  run_timing responsible only for native orchestration
- Clear previously published SDF before rerunning STA so a failed
  rerun cannot expose a stale artifact as current
- Move the run_timing test group into test_module_timing.py with a
  narrow STA-only fake instead of growing the 2.7k-line test_module.py;
  add a success-then-failed-rerun regression test
