+++ test/tools/ecc/test_module.py
@@ -4,8 +4,6 @@ import json
 from pathlib import Path
 from types import SimpleNamespace
 
-import pytest
-
 import chipcompiler.utility as chipcompiler_utility
 from chipcompiler.data import OriginDesign, StepEnum, Workspace
 from chipcompiler.tools.ecc import metrics as ecc_metrics
@@ -414,120 +412,6 @@ def test_ecc_binding_wrappers_stringify_path_arguments(tmp_path):
         ("run_sta", (), {}),
         ("destroy_sta", (), {}),
     ]
-
-
-def test_run_timing_splits_text_reports_and_structured_artifacts(tmp_path):
-    module = ECCToolsModule.__new__(ECCToolsModule)
-    module.ecc = FakeEcc()
-    report_dir = tmp_path / "report" / "MAX_125" / "RCworst"
-    feature_dir = tmp_path / "feature" / "MAX_125" / "RCworst"
-
-    module.run_timing(
-        work_dir=tmp_path / "data" / "sta",
-        report_dir=report_dir,
-        feature_dir=feature_dir,
-        output_modes=("structured", "report"),
-        max_paths_per_analysis=7,
-        corner="MAX_125/RCworst",
-    )
-
-    assert (report_dir / "qor_summary.rpt").is_file()
-    assert (report_dir / "timing_max.rpt").is_file()
-    assert (feature_dir / "qor_summary.json").is_file()
-    assert (feature_dir / "timing_paths.json").is_file()
-    init_config = next(
-        call[1] for call in module.ecc.calls if len(call) == 2 and call[0] == "init_sta"
-    )
-    assert init_config["config_dict"] == {
-        "-temp_directory_path": str(tmp_path / "data" / "sta"),
-        "-output_timing_reports": "1",
-        "-output_timing_features": "1",
-        "-timing_path_limit": "7",
-        "-timing_corner": "MAX_125/RCworst",
-    }
-
-
-def test_run_timing_accepts_qor_summary_without_timing_paths(tmp_path):
-    module = ECCToolsModule.__new__(ECCToolsModule)
-    module.ecc = FakeEcc()
-    module.ecc.structured_timing_filenames = ("qor_summary.json",)
-    feature_dir = tmp_path / "feature" / "MAX_125" / "RCworst"
-
-    module.run_timing(
-        work_dir=tmp_path / "data" / "sta",
-        feature_dir=feature_dir,
-        output_modes=("structured",),
-        corner="MAX_125/RCworst",
-    )
-
-    assert (feature_dir / "qor_summary.json").is_file()
-    assert not (feature_dir / "timing_paths.json").exists()
-
-
-def test_run_timing_rejects_invalid_output_modes(tmp_path):
-    module = ECCToolsModule.__new__(ECCToolsModule)
-    module.ecc = FakeEcc()
-
-    with pytest.raises(ValueError, match="Unsupported STA output modes"):
-        module.run_timing(
-            work_dir=tmp_path / "data" / "sta",
-            feature_dir=tmp_path / "feature",
-            output_modes=("structured", "raw"),
-        )
-
-    assert module.ecc.calls == []
-
-
-def test_run_timing_publishes_sdf_alongside_text_reports(tmp_path):
-    module = ECCToolsModule.__new__(ECCToolsModule)
-    module.ecc = FakeEcc()
-    report_dir = tmp_path / "report" / "MAX_125" / "RCworst"
-
-    module.run_timing(
-        work_dir=tmp_path / "data" / "sta",
-        report_dir=report_dir,
-        output_modes=("report",),
-        corner="MAX_125/RCworst",
-    )
-
-    assert sorted(path.name for path in report_dir.iterdir()) == [
-        "gcd.sdf",
-        "qor_summary.rpt",
-        "timing_max.rpt",
-    ]
-
-
-def test_run_timing_raises_when_sdf_missing(tmp_path):
-    module = ECCToolsModule.__new__(ECCToolsModule)
-    module.ecc = FakeEcc()
-    module.ecc.emit_sdf = False
-    report_dir = tmp_path / "report" / "MAX_125" / "RCworst"
-
-    with pytest.raises(FileNotFoundError, match="SDF"):
-        module.run_timing(
-            work_dir=tmp_path / "data" / "sta",
-            report_dir=report_dir,
-            output_modes=("report",),
-            corner="MAX_125/RCworst",
-        )
-
-    assert not report_dir.exists()
-
-
-def test_run_timing_structured_only_does_not_require_sdf(tmp_path):
-    module = ECCToolsModule.__new__(ECCToolsModule)
-    module.ecc = FakeEcc()
-    module.ecc.emit_sdf = False
-    feature_dir = tmp_path / "feature" / "MAX_125" / "RCworst"
-
-    module.run_timing(
-        work_dir=tmp_path / "data" / "sta",
-        feature_dir=feature_dir,
-        output_modes=("structured",),
-        corner="MAX_125/RCworst",
-    )
-
-    assert (feature_dir / "qor_summary.json").is_file()
 
 
 def test_ecc_runtime_wrappers_stringify_path_arguments(tmp_path):
